Add a whole new set of wrapper objects
This patch adds a wrapper object for Element, Text and Attribute, which
is what React has to offer in terms of node types. Additionally, a
Document type is introduced. However, this is only an internal
abstraction and not available to the end user.

These wrappers gives us somewhat more flexibility. They basically serve
to normalize the internal DOM of React even further. With this, I have
implemented proper support for eg. comparing document order, which was
sort of the last frontier towards full XPath support.