Add a very basic performance test
This consists of a server and a client application that will performing
testing using benchmark.js. To run it, execute the following command

  $ node test/perf/server.js

and navigate to http://localhost:8080 in your favourite browser. You
will see the benchmark results in the console. The implementation is
currently compared against the native implementation and Google's
wgxpath.

The preliminary results show that my implementation is somewhat slower
than Google's.

  document#evaluate x 7,022 ops/sec ±4.02% (41 runs sampled)
  XPathReact#evaluate x 5,268 ops/sec ±5.54% (61 runs sampled)
  Fastest is document#evaluate