Use document.write instead of console.log in perf test
It appears that having the console window open messes greatly with the
performance tests. By writing the results directly to the document, we
can avoid having it open.

The new results shows that using XPathReact on React components is
faster compared to actually rendering them in the DOM and using the
native xpath implementation of the browser. Great success!

  document#evaluate x 7,351 ops/sec ±5.00% (60 runs sampled)
  XPathReact#evaluate x 11,698 ops/sec ±6.41% (61 runs sampled)
  Fastest is XPathReact#evaluate